replace skills to the business card
