+++ src/components/Card/index.tsx
@@ -1,6 +1,7 @@
 import React, { FC, Fragment } from 'react'
 import { useIntl } from 'gatsby-plugin-intl'
 import * as SC from './index.module.scss'
+import Skills from '../Skills'
 import { SiteMetadataTypes } from '../../types'
 import { NekoButton } from 'nekohack-ui'
 import PaperPlaneSvg from '../../static/icons/paper-plane.svg'
@@ -8,27 +9,19 @@ import PaperPlaneSvg from '../../static/icons/paper-plane.svg'
 const Card: FC<any> = ({ data }: SiteMetadataTypes) => {
     const intl = useIntl()
 
-    const positionList: string[] = data.position.split(',')
+    const position: string = data.position.split(',')[
+        Math.floor(Math.random() * 3)
+    ]
 
     return (
         <Fragment>
             <div className={SC.content}>
                 <div className={SC.info}>
                     <div className={SC.me} />
-                    <h1 className={SC.name}>{data.author}</h1>
-                    <div className={SC.position}>
-                        <ul>
-                            {positionList.map((p: string) => {
-                                return <li key={p}>{p}</li>
-                            })}
-                        </ul>
-                    </div>
-                    <p className={SC.location}>
-                        {`${intl.formatMessage({ id: 'locale_is' })} ${
-                            data.lonlat
-                        }`}
-                    </p>
-                    <p className={SC.description}>{data.description}</p>
+                    <h1 className={SC.name}>
+                        {data.author + ' (' + position + ')'}
+                    </h1>
+                    <Skills />
                     <p>
                         <NekoButton
                             onClick={() => {
