+++ src/components/Work/index.module.scss
@@ -32,22 +32,3 @@
         margin: 0 8px;
     }
 }
-
-.skills {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    flex: 0 1 auto;
-    font-size: 0.8rem;
-    line-height: 24px;
-}
-
-.skill {
-    display: flex;
-    flex-direction: column;
-    width: 360px;
-
-    .strong {
-        color: #dd66a1;
-    }
-}
