+++ src/components/Work/index.tsx
@@ -1,16 +1,6 @@
 import React, { FC } from 'react'
 import { useIntl } from 'gatsby-plugin-intl'
 import * as SC from './index.module.scss'
-import { skills } from '../../utils/profile'
-
-type SkillType = {
-    title: string
-    rating: number
-    items: Array<{
-        text: string
-        strong?: boolean
-    }>
-}
 
 type WorkProps = {
     data: GatsbyTypes.Maybe<
@@ -68,32 +58,6 @@ const Work: FC<WorkProps> = ({ data }) => {
                         </a>
                     </li>
                 </ul>
-            </div>
-            <div className={SC.subtitle}>
-                {intl.formatMessage({ id: 'skill' })}
-            </div>
-            <div className={SC.skills}>
-                {skills.map((skill: SkillType, index: number) => {
-                    return (
-                        <div key={index} className={SC.skill}>
-                            <h2>{skill.title}</h2>
-                            {skill.items.map(
-                                (item: { text: string; strong?: boolean }) => {
-                                    return (
-                                        <div
-                                            key={item.text}
-                                            className={
-                                                item.strong ? SC.strong : ''
-                                            }
-                                        >
-                                            {item.text}
-                                        </div>
-                                    )
-                                }
-                            )}
-                        </div>
-                    )
-                })}
             </div>
         </div>
     )
